requirements.txt
Django==3.0.7
pytz==2020.1
sqlparse==0.3.1
asgiref==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.11.0

